requirements.txt
numpy
opencv-python
Pillow
matplotlib
scikit-image
PyQt5
fpdf
